Refactoring: move Instruction to asm module (bpf-profile #7)
